+++ src/main/java/com/LDCream/Byte/proyectoEcommerceLiquidDevs/model/Producto.java
@@ -18,7 +18,7 @@ public class Producto {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_producto")
-    private Integer idProducto;
+    private Long idProducto;
 
     private String nombre;
     private String descripcion;
@@ -35,11 +35,11 @@ public class Producto {
     private Boolean estado;
     public Producto() {}
 
-    public Integer getIdProducto() {
+    public Long getIdProducto() {
         return idProducto;
     }
 
-    public void setIdProducto(Integer idProducto) {
+    public void setIdProducto(Long idProducto) {
         this.idProducto = idProducto;
     }
